Improvements to the Event System
Event listeners no longer need to cast the Event type into the type that is expected, as it is now casted to the correct type by the EventManager.
Created StaticSegment and Line components to make it easier to create and display "ground"-like physics objects.
Cleaned up things in Game.cpp.