Create a countdown timer
Implemented a countdown timer in player view for current demo.

Current total time is set to 30 seconds (as max_countdown in PlayerView.h).

There are 2 styles for the timer representation , which is binary style by default.  It could be set to a regular decimal timer by changing countdown_mode_binary to false.

The current timer is not implemented as an entity. Since there is only one unphysical timer for player and we are using the hybrid event system, I think it might be okay this way for now.